+++ packages/drizzle/src/factory/resolver-postgres.ts
@@ -3,13 +3,13 @@ import {
   type GraphQLFieldOptions,
   type GraphQLSilk,
   type Middleware,
+  MutationFactoryWithResolve,
   type MutationOptions,
   silk,
 } from "@gqloom/core"
 import type { InferSelectModel } from "drizzle-orm"
 import type { PgDatabase, PgTable } from "drizzle-orm/pg-core"
 import type { GraphQLOutputType } from "graphql"
-import { MutationFactoryWithResolve } from "./field"
 import type {
   DeleteArgs,
   InsertArrayArgs,
