+++ packages/drizzle/src/factory/types.ts
@@ -1,4 +1,9 @@
-import type { GraphQLSilk } from "@gqloom/core"
+import type {
+  FieldFactoryWithResolve,
+  GraphQLSilk,
+  MutationFactoryWithResolve,
+  QueryFactoryWithResolve,
+} from "@gqloom/core"
 import type { InferSelectModel, Many, Table } from "drizzle-orm"
 import type { MySqlDatabase } from "drizzle-orm/mysql-core"
 import type { RelationalQueryBuilder as MySqlRelationalQueryBuilder } from "drizzle-orm/mysql-core/query-builders/query"
@@ -6,11 +11,7 @@ import type { PgDatabase } from "drizzle-orm/pg-core"
 import type { RelationalQueryBuilder as PgRelationalQueryBuilder } from "drizzle-orm/pg-core/query-builders/query"
 import type { BaseSQLiteDatabase } from "drizzle-orm/sqlite-core"
 import type { RelationalQueryBuilder as SQLiteRelationalQueryBuilder } from "drizzle-orm/sqlite-core/query-builders/query"
-import type {
-  FieldFactoryWithResolve,
-  MutationFactoryWithResolve,
-  QueryFactoryWithResolve,
-} from "./field"
+
 import type {
   DeleteArgs,
   InsertArrayArgs,
